requirements.txt
python-documentcloud>=4.4.1
requests>=2.31.0
beautifulsoup4>=4.12.0
reportlab>=4.0.0
python-dateutil>=2.8.0
lxml>=4.9.0
atproto>=0.0.46
PyPDF2>=3.0.0